requirements.txt
django>=3.1.1,<3.2.0
djangorestframework>=3.11.1,<3.12.0
graphene-django>=2.15.0,<2.16.0
diff-match-patch
flake8>=4.0.1,<4.1.0